+++ tests/Rest/Integration/MobileIdRestConnectorSessionIT.php
@@ -49,6 +49,7 @@ class MobileIdRestConnectorSessionIT extends TestCase
     {
         $this->connector = MobileIdRestConnector::newBuilder()
             ->withEndpointUrl(TestData::DEMO_HOST_URL)
+            ->withSslPinnedPublicKeys( TestData::DEMO_HOST_PUBLIC_KEY_HASH)
             ->build();
     }